ovl: setxattr: don't deadlock when called from ima_fix_xattr.
IMA tracks the integrity of files by hashing the file content
and storing the hash in an IMA-specific extended attribute
(security.ima). When a file opened for writing is closed by
the last writer, IMA recalculates the hash and updates the
extended attribute. Updating the attribute happens by locking
the inode mutex and calling  __vfs_setxattr_noperm.

For a file on an overlayfs mount, this causes ovl_setxattr
being eventually called (from __vfs_setxattr_noperm via
inode->i_op->setxattr) with the inode already locked. In this
case we cannot do the xattr setting by calling vfs_setxattr
since that will try to lock the inode mutex again. To avoid
a deadlock, we check if the XATTR_IMA_UPDATE flag is set and
call __vfs_setxattr_noperm or vfs_setxattr accordingly.

Upstream-Status: inappropriate [other]

With this patch in place, a deadlock is still possible in
principle as on the offending code path (ima_fix_xattr ->
__ovl_setxattr_noperm -> inode->i_op->setxattr ->
ovl_setxattr) the sb_writer and i_mutex locks are taken in
the wrong order (i_mutex first, then sb_writers). I *think*,
the only way to try and trigger the deadlock would be to modify
an IMA protected file on an overlayfs using a regular write
while at the same time trying to sign the same file with
evmctl ima_sign. It is entirely possible that there other
safeguards present in the code protecting against this corner-
case that I simply missed.

The real fix would be to change IMA, overlayfs, potentially
touching the related parts in VFS as well to eliminate the
the reversed locking. I don't attempt rolling such a patch
myself. I think this needs eyeballing by someone who is already
familiar with VFS, and preferably also with overlayfs.

This bug has been reported upstream to IMA, overlayfs and
linux-fs.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>